Classification/requirements.txt
absl-py==2.4.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
coloredlogs==15.0.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.23.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.9.0
protobuf==3.20.3
pycparser==3.0
PyMatting==1.1.15
pyparsing==3.3.2
python-dateutil==2.9.0.post0
referencing==0.37.0
rembg==2.0.72
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
sounddevice==0.5.5
split-folders==0.6.1
sympy==1.14.0
tifffile==2026.1.28
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
